Rename conditionals -> blackjack
