+++ exercises/concept/blackjack/.docs/introduction.md
@@ -1,14 +1,10 @@
 # Introduction
 
+Conditionals in Go are similar to conditionals in other languages. The underlying type of any conditional operation is the `bool` type, which can have the value of `true` or `false`. Conditionals are often used as flow control mechanisms to check for various conditions.
+
 ## Conditionals If
 
-## Conditionals Switch
-
-## Booleans
-
-Go supports the three logical operators `&&` (AND), `||` (OR), and `!` (NOT).
-
-Conditionals in Go are similar to conditionals in other languages. The underlying type of any conditional operation is the `bool` type, which can have the value of `true` or `false`. Conditionals are often used as flow control mechanisms to check for various conditions. For checking a particular case an `if` statement can be used, which executes its code if the underlying condition is `true` like this:
+For checking a particular case an `if` statement can be used, which executes its code if the underlying condition is `true` like this:
 
 ```go
 var value string
@@ -29,6 +25,8 @@ if value == "val" {
     // default code
 }
 ```
+
+## Conditionals Switch
 
 Go also provides a `switch` statement for more advanced scenarios.
