+++ exercises/concept/blackjack/.meta/exemplar.go
@@ -1,4 +1,4 @@
-package _meta
+package blackjack
 
 // ParseCard returns the integer value of a card following blackjack ruleset.
 func ParseCard(card string) int {
