+++ exercises/concept/blackjack/blackjack.go
@@ -1,4 +1,4 @@
-package conditionals
+package blackjack
 
 // ParseCard returns the integer value of a card following blackjack ruleset.
 func ParseCard(card string) int {
