+++ exercises/concept/blackjack/blackjack_test.go
@@ -1,4 +1,4 @@
-package conditionals
+package blackjack
 
 import "testing"
