+++ exercises/concept/blackjack/go.mod
@@ -1,0 +1,3 @@
+module blackjack
+
+go 1.13
