--- exercises/concept/conditionals/go.mod
@@ -1,3 +1,0 @@
-module conditionals
-
-go 1.13
